fix linting errors

+++ src/crypto/account.rs
@@ -28,7 +28,7 @@ impl Account {
         let mut curve25519_public_key = exchange_keypair.public().id();
 
         unsafe {
-            let account_len = olm_account_size() as usize;
+            let account_len = olm_account_size();
             let account_buf = vec![0_u8; account_len].into_boxed_slice();
             let account = olm_account(Box::into_raw(account_buf) as *mut libc::c_void);
 
@@ -53,7 +53,7 @@ impl Account {
         password: Option<&[u8]>,
     ) -> Result<Account, SelfError> {
         unsafe {
-            let account_len = olm_account_size() as usize;
+            let account_len = olm_account_size();
             let account_buf = vec![0_u8; account_len].into_boxed_slice();
             let account = olm_account(Box::into_raw(account_buf) as *mut libc::c_void);
 
@@ -66,9 +66,9 @@ impl Account {
             olm_unpickle_account(
                 account,
                 password_buf,
-                password_len as u64,
+                password_len,
                 pickle as *mut [u8] as *mut libc::c_void,
-                pickle.len() as u64,
+                pickle.len(),
             );
 
             let account = Account {
@@ -89,7 +89,7 @@ impl Account {
     pub fn one_time_keys(&self) -> Vec<Vec<u8>> {
         unsafe {
             let mut one_time_keys_len = olm_account_one_time_keys_length(self.account);
-            let mut one_time_keys_buf = vec![0_u8; one_time_keys_len as usize].into_boxed_slice();
+            let mut one_time_keys_buf = vec![0_u8; one_time_keys_len].into_boxed_slice();
 
             one_time_keys_len = olm_account_one_time_keys(
                 self.account,
@@ -97,7 +97,7 @@ impl Account {
                 one_time_keys_len,
             );
 
-            let one_time_keys_json = one_time_keys_buf[0..one_time_keys_len as usize].to_vec();
+            let one_time_keys_json = one_time_keys_buf[0..one_time_keys_len].to_vec();
             let one_time_keys: OneTimeKeys =
                 serde_json::from_slice(&one_time_keys_json).expect("always valid json");
 
@@ -112,18 +112,17 @@ impl Account {
 
     pub fn generate_one_time_keys(&mut self, count: usize) -> Result<(), SelfError> {
         unsafe {
-            if (olm_account_max_number_of_one_time_keys(self.account) as usize) < count {
+            if (olm_account_max_number_of_one_time_keys(self.account)) < count {
                 return self.last_error();
             }
 
-            let random_len =
-                olm_account_generate_one_time_keys_random_length(self.account, count as u64);
-            let mut random_buf = vec![0_u8; random_len as usize].into_boxed_slice();
+            let random_len = olm_account_generate_one_time_keys_random_length(self.account, count);
+            let mut random_buf = vec![0_u8; random_len].into_boxed_slice();
             sodium_sys::randombytes_buf(random_buf.as_mut_ptr() as *mut libc::c_void, random_len);
 
             olm_account_generate_one_time_keys(
                 self.account,
-                count as u64,
+                count,
                 random_buf.as_mut_ptr() as *mut libc::c_void,
                 random_len,
             );
@@ -150,7 +149,7 @@ impl Account {
     pub fn identity_keys(&self) -> Vec<u8> {
         unsafe {
             let mut identity_keys_len = olm_account_identity_keys_length(self.account);
-            let mut identity_keys_buf = vec![0_u8; identity_keys_len as usize].into_boxed_slice();
+            let mut identity_keys_buf = vec![0_u8; identity_keys_len].into_boxed_slice();
 
             identity_keys_len = olm_account_identity_keys(
                 self.account,
@@ -158,14 +157,14 @@ impl Account {
                 identity_keys_len,
             );
 
-            identity_keys_buf[0..identity_keys_len as usize].to_vec()
+            identity_keys_buf[0..identity_keys_len].to_vec()
         }
     }
 
     pub fn pickle(&self, password: Option<&[u8]>) -> Result<Vec<u8>, SelfError> {
         unsafe {
             let mut account_pickle_len = olm_pickle_account_length(self.account);
-            let mut account_pickle_buf = vec![0_u8; account_pickle_len as usize].into_boxed_slice();
+            let mut account_pickle_buf = vec![0_u8; account_pickle_len].into_boxed_slice();
 
             let password_len = password.map(|pwd| pwd.len()).unwrap_or(0);
 
@@ -176,14 +175,14 @@ impl Account {
             account_pickle_len = olm_pickle_account(
                 self.account,
                 password_buf,
-                password_len as u64,
+                password_len,
                 account_pickle_buf.as_mut_ptr() as *mut libc::c_void,
                 account_pickle_len,
             );
 
             self.last_error()?;
 
-            Ok(account_pickle_buf[0..account_pickle_len as usize].to_vec())
+            Ok(account_pickle_buf[0..account_pickle_len].to_vec())
         }
     }
 
@@ -208,9 +207,9 @@ impl Account {
                 session.as_mut_ptr(),
                 self.account,
                 identity_key_buf.as_ptr() as *const libc::c_void,
-                identity_key_buf.len() as u64,
+                identity_key_buf.len(),
                 one_time_message_buf.as_mut_ptr() as *mut libc::c_void,
-                one_time_message_buf.len() as u64,
+                one_time_message_buf.len(),
             );
 
             session.last_error()?;
@@ -235,16 +234,16 @@ impl Account {
 
         unsafe {
             let random_len = olm_create_outbound_session_random_length(session.as_mut_ptr());
-            let mut random_buf = vec![0_u8; random_len as usize].into_boxed_slice();
+            let mut random_buf = vec![0_u8; random_len].into_boxed_slice();
             sodium_sys::randombytes_buf(random_buf.as_mut_ptr() as *mut libc::c_void, random_len);
 
             olm_create_outbound_session(
                 session.as_mut_ptr(),
                 self.account,
                 identity_key_buf.as_ptr() as *const libc::c_void,
-                identity_key_buf.len() as u64,
+                identity_key_buf.len(),
                 one_time_key.as_ptr() as *const libc::c_void,
-                one_time_key.len() as u64,
+                one_time_key.len(),
                 random_buf.as_ptr() as *mut libc::c_void,
                 random_len,
             );
